Extract message generating function
